pyproject.toml
[tool.poetry]
name = "dataflow-composer-migration"
version = "0.1.0"
description = "Tool to migrate Apache Beam pipelines from Google Cloud Dataflow to Cloud Composer"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-dataflow-client = "^0.8.3"
google-cloud-storage = "^2.14.0"
jinja2 = "^3.1.3"
javalang = "^0.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
transform = "scripts.beam_code_transformer:main"
validate = "scripts.beam_code_validator:main"
fix = "scripts.beam_code_fixer:main"
migrate = "scripts.dataflow_migration:main"
